BookingFilter class implemented
