+++ Browser Automation/Barcelo Booking Bot/main.py
@@ -4,11 +4,9 @@ import sys
 
 with Booking(teardown=False) as bot:
     bot.land_first_page()
-    bot.accept_cookies()
 
-    bot.sort_cheapest()
-    bot.choose_location()
-    bot.filter_brand()
+    bot.apply_filters()
+    
     #time.sleep(3) #Let page load after filters
     bot.report(int(sys.argv[1]), int(sys.argv[2]))
     #time.sleep(5)
